requirements.txt
# Eva Live - AI Virtual Presenter System
# Core dependencies for running Eva Live

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# AI and ML libraries
openai>=1.3.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0

# Vector database and caching
pinecone-client>=2.2.0
redis>=5.0.0

# Audio processing
azure-cognitiveservices-speech>=1.34.0
pydub>=0.25.1
soundfile>=0.12.1
librosa>=0.10.1
noisereduce>=3.0.0
pygame>=2.5.2

# Image and video processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
imageio>=2.31.0

# Document processing
python-pptx>=0.6.21
PyPDF2>=3.0.1
python-docx>=0.8.11
markdown>=3.5.0

# HTTP and async
aiohttp>=3.9.0
httpx>=0.25.0
asyncio-mqtt>=0.13.0

# Data processing
pandas>=2.1.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
tqdm>=4.66.0

# Optional: Enhanced features
# scikit-learn>=1.3.0  # For advanced NLP
# matplotlib>=3.7.0    # For visualizations
# gradio>=3.50.0       # For web UI
